# Improve diff performance