Use a map of strings so we can do O(1) lookups. Now O(n) instead of O(n^2).

Tested via benchmark in lemming.

before:
```
BenchmarkGNMISet/config_replace-12                     2        1360487239 ns/op
BenchmarkGNMISet/config_update-12                      1        7883432860 ns/op
BenchmarkGNMISet/config_replace_and_delete-12                  1        1994916012 ns/op
```

after:
```
BenchmarkGNMISet/config_replace_nonleaf-12                     2         824160406 ns/op
BenchmarkGNMISet/config_update_leaf-12                         1        4868570333 ns/op
BenchmarkGNMISet/config_replace_and_delete-12                  1        1420574251 ns/op
```